requirements.txt
asgiref==3.4.1
Django==3.2.7
django-froala-editor==3.2.2
django-quill-editor==0.1.22
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
